Option to manage Infra for users that do not have admin of org policies (otherwise setup script fails)

+++ setup.sh
@@ -87,33 +87,35 @@ while [ -z "$enabled" ]; do
   sleep 5;
 done
 
-$printf "Setting Policy Constraints..."  | tee -a "$LOG"
-gcloud org-policies reset constraints/iam.disableServiceAccountCreation --project=$PROJECT_ID
-gcloud org-policies reset constraints/iam.disableServiceAccountKeyCreation --project=$PROJECT_ID
-gcloud org-policies reset constraints/compute.vmExternalIpAccess --project=$PROJECT_ID
-gcloud org-policies reset constraints/compute.requireShieldedVm --project=$PROJECT_ID
-gcloud org-policies reset constraints/storage.restrictAuthTypes --project=$PROJECT_ID
+if [ -n "$DISABLE_POLICY" ]; then
+  $printf "Setting Policy Constraints..."  | tee -a "$LOG"
+  gcloud org-policies reset constraints/iam.disableServiceAccountCreation --project=$PROJECT_ID
+  gcloud org-policies reset constraints/iam.disableServiceAccountKeyCreation --project=$PROJECT_ID
+  gcloud org-policies reset constraints/compute.vmExternalIpAccess --project=$PROJECT_ID
+  gcloud org-policies reset constraints/compute.requireShieldedVm --project=$PROJECT_ID
+  gcloud org-policies reset constraints/storage.restrictAuthTypes --project=$PROJECT_ID
 
-echo "Allow up to 30 seconds to propagate the policy changes..."
-sleep 30
-echo "Policy Changes done"
-# Otherwise fails on PreconditionException: 412 Request violates constraint 'constraints/iam.disableServiceAccountKeyCreation'
+  echo "Allow up to 30 seconds to propagate the policy changes..."
+  sleep 30
+  echo "Policy Changes done"
+  # Otherwise fails on PreconditionException: 412 Request violates constraint 'constraints/iam.disableServiceAccountKeyCreation'
 
-ready=$(gcloud org-policies describe constraints/iam.disableServiceAccountKeyCreation --project=$PROJECT_ID 2>/dev/null)
-while [ -z "$ready" ]; do
   ready=$(gcloud org-policies describe constraints/iam.disableServiceAccountKeyCreation --project=$PROJECT_ID 2>/dev/null)
-  sleep 5;
-done
+  while [ -z "$ready" ]; do
+    ready=$(gcloud org-policies describe constraints/iam.disableServiceAccountKeyCreation --project=$PROJECT_ID 2>/dev/null)
+    sleep 5;
+  done
 
-gcloud resource-manager org-policies describe    compute.trustedImageProjects --project=$PROJECT_ID    --effective > policy.yaml
-echo "  - projects/illumina-dragen" >> policy.yaml
-echo "  - projects/$GCLOUD_IMAGE_PROJECT" >> policy.yaml
-echo "  - projects/$GCLOUD_BATCH_IMAGE_PROJECT" >> policy.yaml
-gcloud resource-manager org-policies set-policy \
-   policy.yaml --project=$PROJECT_ID
-rm  policy.yaml
+  gcloud resource-manager org-policies describe    compute.trustedImageProjects --project=$PROJECT_ID    --effective > policy.yaml
+  echo "  - projects/illumina-dragen" >> policy.yaml
+  echo "  - projects/$GCLOUD_IMAGE_PROJECT" >> policy.yaml
+  echo "  - projects/$GCLOUD_BATCH_IMAGE_PROJECT" >> policy.yaml
+  gcloud resource-manager org-policies set-policy \
+     policy.yaml --project=$PROJECT_ID
+  rm  policy.yaml
 
-$printf "Finished with Org policy Update"  INFO | tee -a "$LOG"
+  $printf "Finished with Org policy Update"  INFO | tee -a "$LOG"
+fi
 
 network=$(gcloud compute networks list --filter="name=(\"$GCLOUD_NETWORK\" )" --format='get(NAME)' 2>/dev/null)
 if [ -z "$network" ]; then
